Add return types to the FromXmlDoc and FromXmlNode extension methods
While trying to add test fields for nested types, noticed that the
compiler was freaking out because the `Node`, `NodeOpt`, etc. types in
`source_types.fsx` all have the same field name, so it could not decide
which type was being returned from calls to FromXmlNode and FromXmlDoc.
Good thing I decided to write tests?