Adjust handling of XmlNode on typedefs to be case insensitive
Adjusted the `thingFromDocElement` helper to go through the
`XmlDocument.ChildNodes` property from the given document, rather than
calling `GetElementsByTagName`. To continue resolving to the correct
elements, added a filter against the element name using the
`String.Equals` member, and specified
`StringComparison.OrdinalIgnoreCase` to make this case insensitive. Of
course, the XPath helpers cannot be adjusted in this way, but one can
dream.