Refactor FromXml helper functions
After a long night of staring at the ceiling, pacing, and excessive
whiteboard use, figured out a decent set of functions that can be
combined appropriately to meet all of the current match cases.

The new "getThingValues" functions take the current XML node and a
target tag/attribute name or XPath specifier and a parser function for
the desired type, then return all matching results as a sequence.

For each field, the appropriate getThingValues function can then be
passed as a "getter" to maybeOne, exactlyOne, or one of the
getCollection or getMaybeCollection functions, where "Collection" is
one of Array, List, or Seq. These functions pass values along to the
specified "getter" and process the resulting sequence, either using
Seq.tryHead to get the first matching result and doing something with
it, or using Seq.toXYZ to convert all results to the desired collection
type.

As previously done, any optional collections are set to None if zero
matching elements were processed, and any fields using the `seq`
collection type are cached as an array before being converted back into
a `seq` to prevent exceptions from being thrown from a "parser"
function while someone else is trying to use the parsed results. No
lazy-loading here.

Lastly, two helper functions are added to help facilitate creating a
"parser". The `getInnerText` function simply returns the inner text of
the given XmlNode, used for string fields, and the `parserForStrings`
function (likely to be renamed at some point) takes a function that
returns a boolean and a result and converts it into a function that
"always" returns a result, and throws an exception if the value could
not be parsed; the latter function is for use with TryParse-style
functions.