Adjust FromXml helper `parserForStrings` to take another parameter
Updated the expansions that use it as well; some code lines are about to
get a bit longer? Alternative plan would be to look up the stack and
find the closest `FromXmlNode` or `FromXmlDoc` method call and figure
out the target type from that, although the specific field name would
still need to be planted like this.