[FromXml] Add `parserForSimpleUnion` helper
Takes a target type and field name and returns a function which takes an
`XmlNode` and processes it into a `'T` value, where `'T` is a union type which
contains a case whose name roughly matches the inner text of the `XmlNode`.
As elsewhere, name matching is not case-sensitive. Union cases that are to be
processed cannot contain fields.