[IVM] Add a new ImmutableViewModel template
This is a template I have been developing on-and-off for quite some
time, and finally made some useful progress on. The base idea, which
will be documented in the README shortly, is to take a user-defined data
model, and generate a WPF-friendly view model. The data model, intended
to be defined as an immutable F# record type, is then used by the view
model to maintain its internal state. Functions defined in a nearby
module are repurposed to build `ICommand` instances, allowing additional
parameters to be passed, command disabling based on current state, and
asynchronous command behaviors.