[IVM] Adjust nested lists to use both MakeObservableList overloads
For nested types with a key field, this should improve UI behavior.

While changing this, also replaced the use of `getTypeNameWithViewModel`
with `getTypeName`, so that nested types will not use their view model
type in a list. This should be properly supported later, when I get a
chance to write up another overload for `MakeObservableList`.